refactor: fossilBase.ts -> openedRepository.ts
